refactor(github): split RepoAnalysisForm into custom hooks
- Extract token management, repo URLs management, and repo analysis logic into separate hooks
- Update CommitsTab component to reflect new commit data limitations
- Refactor index.ts to re-export all hooks
- Create new files for GitHub-specific hooks:
- github/index.ts
  - github/useRepoAnalysis.ts
  - github/useRepoUrlsManagement.ts
  - github/useTokenManagement.ts
