+++ src/hooks/index.ts
@@ -1,3 +1,7 @@
-export { useCollapseStore } from './useCollapseStore';
-export { useGitHubToken, type TokenMessage } from './useGitHubToken';
-export { useRepoActions } from './useRepoActions'; 
+// Re-export all hooks
+export * from './useCollapseStore';
+export * from './useGitHubToken';
+export * from './useRepoActions';
+
+// GitHub specific hooks
+export * from './github';
